fix(admin/broadcasts): 할인율 소수점(83.3 등) 입력 시 저장 실패 해결
discountRate zod 스키마의 .int() 제약으로 소수점 값이 들어오면 400 응답이
떨어져 수정 저장이 실패하던 문제. 할인율은 소수점이 자연스러우므로 정수
제한을 제거하고 0~100 범위만 유지.

+++ src/app/api/admin/broadcasts/route.ts
@@ -35,7 +35,7 @@ const baseSchema = z.object({
   productIds: z.array(z.string()).default([]),
   specialPrice: z.coerce.number().int().min(0).optional().nullable(),
   regularPrice: z.coerce.number().int().min(0).optional().nullable(),
-  discountRate: z.coerce.number().int().min(0).max(100).optional().nullable(),
+  discountRate: z.coerce.number().min(0).max(100).optional().nullable(),
   livestreamUrl: z.string().url('올바른 URL을 입력하세요.').or(z.literal('')).optional().nullable(),
   vodUrl: z.string().url('올바른 URL을 입력하세요.').or(z.literal('')).optional().nullable(),
   description: z.string().max(2000).optional().nullable(),
